requirements.txt
Flask==2.2.2
PyJWT==2.4.0
Flask-Limiter==2.0.4
python-dotenv==1.0.0
Flask-Cors==3.0.10
gunicorn==20.1.0
